+++ simulator/src/sensors/mod.rs
@@ -1,12 +1,12 @@
-use simulator_types::PhysicsState;
-
 pub mod barometer;
 pub mod contact;
 pub mod engine_feedback;
 pub mod imu;
 pub mod lidar;
 
-/// A sensor reads the current world snapshot and returns a vector of scalar observations.
+use simulator_types::PhysicsState;
+
+/// A sensor reads the current world snapshot and returns scalar observations.
 pub trait Sensor: Send + Sync {
-    fn observe(&self, state: &PhysicsState) -> Vec<f64>;
+    fn observe(&self, state: &PhysicsState) -> Vec<f32>;
 }
